add helmet and refactor server
