+++ server/utils/database.js
@@ -8,7 +8,8 @@ const connectDB = async () => {
 
     console.log(`MongoDB Connected: ${conn.connection.host}`);
   } catch (err) {
-    console.error(`Error: ${err.message}`);
+    console.error(`MongoDB Connection Error: ${err.message}`);
+    console.log(err);
     process.exit(1);
   }
 };
